Mods to status data info, nagiostats for tracking buffer usage, minor patches to avail and status CGIs
git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/branches/nagios-2-x-bugfixes@793 5f96b256-904b-4d8d-8c98-d829582c6739

+++ base/commands.c
@@ -3,7 +3,7 @@
  * COMMANDS.C - External command functions for Nagios
  *
  * Copyright (c) 1999-2005 Ethan Galstad (nagios@nagios.org)
- * Last Modified:   12-21-2005
+ * Last Modified:   12-26-2005
  *
  * License:
  *
@@ -85,7 +85,7 @@ passive_check_result    *passive_check_result_list_tail=NULL;
 
 extern pthread_t       worker_threads[TOTAL_WORKER_THREADS];
 extern circular_buffer external_command_buffer;
-extern unsigned long   external_command_buffer_slots;
+extern int             external_command_buffer_slots;
